app/requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
flask==3.1.0
flask-cors==5.0.1
flask-session==0.8.0
gcloud==0.18.3
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jws==0.1.3
markupsafe==3.0.2
msgspec==0.19.0
oauth2client==4.1.3
packaging==25.0
pip==25.0.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycryptodome==3.22.0
pyparsing==3.2.3
pyrebase4==4.8.0
python-dotenv==1.1.0
python-jwt==2.0.1
redis==6.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==79.0.0
six==1.17.0
urllib3==1.26.20
uv==0.6.16
watchdog==6.0.0
werkzeug==3.1.3
wtforms==3.2.1
